export lazy current config
