+++ packages/cli/src/error.rs
@@ -81,6 +81,12 @@ impl From<dioxus_cli_config::CargoError> for Error {
     }
 }
 
+impl From<dioxus_cli_config::CrateConfigError> for Error {
+    fn from(e: dioxus_cli_config::CrateConfigError) -> Self {
+        Self::RuntimeError(e.to_string())
+    }
+}
+
 #[macro_export]
 macro_rules! custom_error {
     ($msg:literal $(,)?) => {
