+++ packages/cli/src/lib.rs
@@ -13,9 +13,6 @@ pub use builder::*;
 pub mod cli;
 pub use cli::*;
 
-pub mod config;
-pub use config::*;
-
 pub mod error;
 pub use error::*;
